Reduce code size for non-FP config
The non-FP config was still handling float formats but would print out a
'?' instead of returning EXBADFORMAT. This change removes that and turns
float conversions including fixed point into bad formats. The idea being
that platforms that turn off FP support would not use them in the first
place, and so reporting a bad format is more suitable.